+++ dotnet/Eyes.Appium.DotNet/Fluent/IAppiumCheckTarget.cs
@@ -4,5 +4,6 @@
     {
         bool? IsDefaultWebview();
         string GetWebview();
+        bool? GetScreenshotMode();
     }
 }