+++ dotnet/Eyes.Appium2.DotNet/Eyes.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Globalization;
 using Applitools.Fluent;
 using Applitools.Utils;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Appium;
-using OpenQA.Selenium.Remote;
 
 namespace Applitools.Appium
 {
@@ -12,10 +10,6 @@ namespace Applitools.Appium
     {
         private readonly Selenium.Eyes seleniumEyes_;
         private IWebDriver driver_;
-
-        //public IDictionary<string, object> CachedSessionDetails { get; private set; }
-
-        //public static Func<IWebDriver, bool> IsLandscapeOrientation = IsLandscapeOrientationImpl_;
 
         #region Constructors
 
@@ -253,6 +247,7 @@ namespace Applitools.Appium
             var request = seleniumEyes_.CreateCheckRequest(checkSettings);
             request.Payload.Settings.Webview = (object) appiumCheckTarget.GetWebview() ??
                                               appiumCheckTarget.IsDefaultWebview();
+            request.Payload.Settings.ScreenshotMode = ToScreenshotMode_(appiumCheckTarget.GetScreenshotMode());
             seleniumEyes_.SendCheckRequest(request);
         }
 
@@ -295,6 +290,14 @@ namespace Applitools.Appium
         public override TestResults Close(bool throwEx = true)
         {
             return seleniumEyes_.Close(throwEx);
+        }
+        
+        private static string ToScreenshotMode_(bool? screenshotMode) {
+            if (screenshotMode == null) {
+                return null;
+            }
+
+            return screenshotMode.Value ? "default" : "applitools-lib";
         }
     }
 }