+++ dotnet/Eyes.Selenium.DotNet/Eyes.cs
@@ -594,13 +594,6 @@ namespace Applitools.Selenium
                 {
                     Eyes = eyes_,
                     Config = eyesConfig,
-                    // Target = new DriverTarget
-                    // {
-                    //     ServerUrl = serverUrl,
-                    //     Capabilities = GetCapabilities(remoteWebDriver),
-                    //     SessionId = remoteWebDriver.SessionId.ToString(),
-                    //     Proxy = Config.Proxy
-                    // },
                     Target = null,
                     Settings = CreateUniversalCheckSettings(seleniumCheckSettings)
                 }